fix(droid): hooks.json 改写 flat 顶层并映射 Stop→TurnCompleted
- 独立 hooks.json 事件键直接在顶层（官方规范）；历史 wrapped 结构被
  droid 整份忽略、hook 永不执行。新增 hooksKeyStyle: "flat"，安装时
  迁移清理旧 wrapped Pier 条目，用户条目原样保留
- fallback settings.json 用户 hooks 合并进 flat 顶层事件键（droid 唯一
  读取形态），不再写进死的 wrapped 记录
- 官方 Stop 只在主 agent 完成回复时发射（取消只发 Notification）→
  直接映射 TurnCompleted（copilot agentStop 同款），ready/completed 可达
- PIER_HOOK_COMMAND_GENERATION 16→17 触发存量重装
- 证据矩阵与轨迹 fixture 同步：ready/completed 入 native，覆盖 156→158
- shared.ts 超 500 行硬上限，拆出 hooks/nested.ts（嵌套工厂）与
  hooks/config.ts（配置 IO / 世代守卫），shared 转纯兼容再导出
