+++ docs/superpowers/specs/2026-07-13-agent-status-adapter-contract-audit.md
@@ -118,7 +118,7 @@ Grok：`turn_completed` cancelled/end_turn。
 | crush | 历史配置清理器 | `coarse` | 无 | `none` | 无 |
 | cursor | Cursor hook 配置 | `full` | `sessionStart→SessionStart`；`beforeSubmitPrompt→PromptSubmit`；`preToolUse→ToolStart/InteractionRequested/SubagentStart`；`postToolUse→ToolComplete/InteractionResolved/SubagentStop`；`postToolUseFailure→ToolComplete/InteractionResolved/SubagentStop`；`subagentStart→SubagentStart`；`subagentStop→SubagentStop`；`stop.status=completed→TurnCompleted`；`stop.status=aborted→TurnInterrupted`；`stop.status=error→error`；`sessionEnd→SessionEnd` | `advisory` | 无 |
 | devin | JSON hook 配置 | `full` | `SessionStart→SessionStart`；`UserPromptSubmit→PromptSubmit`；`Stop→Stop`；`PostCompaction→processing`；`SessionEnd→SessionEnd`；`PreToolUse→ToolStart`；`PostToolUse→ToolComplete` | `advisory` | 无 |
-| droid | 嵌套 JSON hook | `full` | `SessionStart/SessionEnd→同名`；`UserPromptSubmit→PromptSubmit`；`Stop→Stop`；`Notification→TurnInterrupted/processing`；`PreCompact→processing`；`PreToolUse→ToolStart`；`PostToolUse→ToolComplete` | `advisory` | 无 |
+| droid | 嵌套 JSON hook | `full` | `SessionStart/SessionEnd→同名`；`UserPromptSubmit→PromptSubmit`；`Stop→TurnCompleted`；`Notification→TurnInterrupted/processing`；`PreCompact→processing`；`PreToolUse→ToolStart`；`PostToolUse→ToolComplete` | `advisory` | 无 |
 | gemini | Gemini hook 配置 | `full` | `SessionStart/SessionEnd→同名`；`BeforeAgent→PromptSubmit`；`AfterAgent→Stop`；`PreCompress→processing`；`BeforeTool→ToolStart`；`AfterTool→ToolComplete` | `advisory` | 无 |
 | goose | 受管理的 Goose 插件 hook | `full` | `SessionStart/SessionEnd→同名`；`UserPromptSubmit→PromptSubmit`；`PreToolUse→ToolStart`；`PostToolUse/PostToolUseFailure→ToolComplete`；`Stop→Stop` | `advisory` | 无 |
 | grok | 嵌套 JSON hook | `full` | `SessionStart/SessionEnd→同名`；`UserPromptSubmit→PromptSubmit`；`PreToolUse→ToolStart/InteractionRequested`；`PostToolUse→ToolComplete/InteractionResolved`；`PostToolUseFailure→ToolComplete/InteractionResolved`；`PermissionDenied→ToolComplete/InteractionResolved`；`Stop→Stop`；`StopCancelled→TurnInterrupted`；`StopFailure→error`；`SubagentStart/SubagentStop→同名`；`PreCompact/PostCompact→processing` | `advisory` | 无 |
